porting more tests
