+++ tests/scenarios/engines-test.ts
@@ -3,7 +3,10 @@ import { PreparedApp, Project } from 'scenario-tester';
 import QUnit from 'qunit';
 import merge from 'lodash/merge';
 import { setupFastboot, loadFromFixtureData } from './helpers';
-import { dirname } from 'path';
+import { ExpectFile, expectFilesAt } from '@embroider/test-support';
+import { dirname, join } from 'path';
+import { readFileSync } from 'fs';
+
 const { module: Qmodule, test } = QUnit;
 
 // Both ember-engines and its dependency ember-asset-loader have undeclared
@@ -30,7 +33,7 @@ function emberEngines(): Project {
   return engines;
 }
 
-appScenarios
+let engineScenarios = appScenarios
   .only('lts_3_28') // ember-engines doesn't have an ember 4.0 compatible release yet.
   .map('engines', project => {
     let eagerEngine = baseAddon();
@@ -74,13 +77,50 @@ appScenarios
 
     let engineTestFiles = loadFromFixtureData('engines-host-app');
     merge(project.files, engineTestFiles);
-  })
-  .expand({
-    'with-fastboot': project => {
-      project.linkDependency('ember-cli-fastboot', { baseDir: __dirname });
-      project.linkDependency('fastboot', { baseDir: __dirname });
-    },
-    'without-fastboot': () => {},
+  });
+
+engineScenarios
+  .map('without-fastboot', () => {})
+  .forEachScenario(scenario => {
+    Qmodule(scenario.name, function (hooks) {
+      let app: PreparedApp;
+      let expectFile: ExpectFile;
+
+      hooks.before(async assert => {
+        app = await scenario.prepare();
+        let result = await app.execute('yarn run build', { env: { STAGE2_ONLY: 'true' } });
+        assert.equal(result.exitCode, 0, result.output);
+      });
+
+      hooks.beforeEach(assert => {
+        expectFile = expectFilesAt(readFileSync(join(app.dir, 'dist/.stage2-output'), 'utf8'), { qunit: assert });
+      });
+
+      test(`yarn test`, async function (assert) {
+        let result = await app.execute('yarn test');
+        assert.equal(result.exitCode, 0, result.output);
+      });
+
+      test('lazy engines appear in _embroiderEngineBundles_', function () {
+        expectFile('assets/app-template.js').matches(/import\("\.\/_engine_\/lazy-engine\.js"\)/);
+      });
+
+      test('lazy engine css is imported', function () {
+        expectFile('assets/_engine_/lazy-engine.js').matches(
+          /i\("\.\.\/\.\.\/node_modules\/lazy-engine\/lazy-engine\.css"\)/
+        );
+      });
+
+      test('eager engine css is merged with vendor.css', function () {
+        expectFile('assets/vendor.css').matches(`.eager { background-color: blue; }`);
+      });
+    });
+  });
+
+engineScenarios
+  .map('with-fastboot', app => {
+    app.linkDependency('ember-cli-fastboot', { baseDir: __dirname });
+    app.linkDependency('fastboot', { baseDir: __dirname });
   })
   .forEachScenario(scenario => {
     Qmodule(scenario.name, function (hooks) {
@@ -94,37 +134,27 @@ appScenarios
         assert.equal(result.exitCode, 0, result.output);
       });
 
-      if (/with-fastboot/.test(scenario.name)) {
-        Qmodule(`fastboot`, function (hooks) {
-          let visit: any;
+      let visit: any;
 
-          hooks.before(async () => {
-            ({ visit } = await setupFastboot(app));
-          });
+      hooks.before(async () => {
+        ({ visit } = await setupFastboot(app));
+      });
 
-          test('host-app', async function (assert) {
-            let doc = (await visit('/')).window.document;
-            assert.equal(
-              doc.querySelector('[data-test-duplicated-helper]').textContent.trim(),
-              'from-engines-host-app'
-            );
-          });
+      test('host-app', async function (assert) {
+        let doc = (await visit('/')).window.document;
+        assert.equal(doc.querySelector('[data-test-duplicated-helper]').textContent.trim(), 'from-engines-host-app');
+      });
 
-          test('lazy-engine', async function (assert) {
-            let doc = (await visit('/use-lazy-engine')).window.document;
-            assert.equal(doc.querySelector('[data-test-lazy-engine-main] > h1').textContent.trim(), 'Lazy engine');
-            assert.equal(doc.querySelector('[data-test-duplicated-helper]').textContent.trim(), 'from-lazy-engine');
-          });
+      test('lazy-engine', async function (assert) {
+        let doc = (await visit('/use-lazy-engine')).window.document;
+        assert.equal(doc.querySelector('[data-test-lazy-engine-main] > h1').textContent.trim(), 'Lazy engine');
+        assert.equal(doc.querySelector('[data-test-duplicated-helper]').textContent.trim(), 'from-lazy-engine');
+      });
 
-          test('eager-engine', async function (assert) {
-            let doc = (await visit('/use-eager-engine')).window.document;
-            assert.equal(doc.querySelector('[data-test-eager-engine-main] > h1').textContent.trim(), 'Eager engine');
-            assert.equal(
-              doc.querySelector('[data-test-duplicated-helper]').textContent.trim(),
-              'from-eager-engine-helper'
-            );
-          });
-        });
-      }
+      test('eager-engine', async function (assert) {
+        let doc = (await visit('/use-eager-engine')).window.document;
+        assert.equal(doc.querySelector('[data-test-eager-engine-main] > h1').textContent.trim(), 'Eager engine');
+        assert.equal(doc.querySelector('[data-test-duplicated-helper]').textContent.trim(), 'from-eager-engine-helper');
+      });
     });
   });
